requirements.txt
alabaster>=0.7.12
Babel>=2.8.0
bleach>=3.3.0
certifi>=2020.6.20
cffi>=1.14.1
chardet>=3.0.4
cloudpickle>=1.3.0
colorama>=0.4.3
cryptography>=3.3.2
cycler>=0.10.0
decorator>=4.4.2
docutils>=0.16
future>=0.18.2
gym>=0.17.2
idna>=2.10
imagesize>=1.2.0
importlib-metadata>=1.7.0
jeepney>=0.4.3
Jinja2>=2.11.3
keyring>=21.2.1
kiwisolver>=1.2.0
MarkupSafe>=1.1.1
matplotlib>=3.3.0
networkx>=2.4
numpy>=1.19.1
packaging>=20.4
Pillow>=8.2.0
pkginfo>=1.5.0.1
prettytable>=0.7.2
pycparser>=2.20
pyglet>=1.5.0
Pygments>=2.7.4
pyparsing>=2.4.7
python-dateutil>=2.8.1
pytz>=2020.1
PyYAML>=5.4
readme-renderer>=26.0
requests>=2.24.0
requests-toolbelt>=0.9.1
rfc3986>=1.4.0
scipy>=1.5.2
SecretStorage>=3.1.2
six>=1.15.0
snowballstemmer>=2.0.0
Sphinx>=3.1.2
sphinx-rtd-theme>=0.5.0
sphinxcontrib-applehelp>=1.0.2
sphinxcontrib-devhelp>=1.0.2
sphinxcontrib-htmlhelp>=1.0.3
sphinxcontrib-jsmath>=1.0.1
sphinxcontrib-qthelp>=1.0.3
sphinxcontrib-serializinghtml>=1.1.4
tqdm>=4.48.0
twine>=3.2.0
urllib3
webencodings>=0.5.1
zipp>=3.1.0
